+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Display child index pages on rendered Subsite home pages
 - Synchronise docker-compose file version numbers
 - Add list_creatable_pages management command
+- Update parent_page_types and subpage_types attrs for many models
 
 ## 63.1 (2022-10-03)